refactor bounties
